+++ src/main/java/org/realityforge/replicant/client/json/gwt/ReplicantConfig.java
@@ -10,4 +10,6 @@ public interface ReplicantConfig
   boolean shouldValidateRepositoryOnLoad();
 
   boolean repositoryDebugOutputEnabled();
+
+  boolean subscriptionsDebugOutputEnabled();
 }